他でも使うクラスなのでindexに移動
